Refactor primitive value generations
Move the responsibility to generate primitive values from
org.javaunit.autoparams package to org.javaunit.autoparams.generator
package.
