+++ autoparams/src/main/java/org/javaunit/autoparams/Factories.java
@@ -11,22 +11,12 @@ import java.util.stream.DoubleStream;
 import java.util.stream.IntStream;
 import java.util.stream.LongStream;
 
+@Deprecated
+@SuppressWarnings("DeprecatedIsStillUsed")
 final class Factories {
 
     private static ThreadLocalRandom random() {
         return ThreadLocalRandom.current();
-    }
-
-    public static boolean createBoolean() {
-        return random().nextInt() % 2 == 0;
-    }
-
-    public static byte createByte() {
-        return (byte) random().nextInt(Byte.MIN_VALUE, Byte.MAX_VALUE + 1);
-    }
-
-    public static short createShort() {
-        return (short) random().nextInt(Short.MIN_VALUE, Short.MAX_VALUE + 1);
     }
 
     public static int createInt() {
@@ -37,16 +27,8 @@ final class Factories {
         return random().nextLong();
     }
 
-    public static float createFloat() {
-        return random().nextFloat();
-    }
-
     public static double createDouble() {
         return random().nextDouble();
-    }
-
-    public static char createChar() {
-        return (char) random().nextInt(Character.MAX_VALUE + 1);
     }
 
     public static BigInteger createBigInteger() {
@@ -97,3 +79,4 @@ final class Factories {
     }
 
 }
+
